+++ src/aat/extended_kerning.rs
@@ -1,11 +1,11 @@
 use core::convert::TryFrom;
 
-use ttf_parser::ankr;
+use ttf_parser::{aat, ankr};
+use ttf_parser::parser::FromData;
 
 use crate::Face;
 use crate::buffer::{BufferScratchFlags, Buffer};
 use crate::ot::{attach_type, lookup_flags, ApplyContext, TableIndex};
-use crate::tables::aat;
 use crate::tables::kerx;
 use crate::plan::ShapePlan;
 use crate::ot::matching::SkippyIter;
@@ -154,16 +154,51 @@ fn apply_simple_kerning(
     }
 }
 
-// TODO: use crate::tables::aat instead
+const START_OF_TEXT: u16 = 0;
 
-fn apply_state_machine_kerning<E: aat::Entry, T: aat::StateTable2<E>>(
+trait KerxEntryDataExt {
+    fn new_state(&self) -> u16;
+    fn flags(&self) -> u16;
+
+    /// If set, reset the kerning data (clear the stack).
+    fn has_reset(&self) -> bool;
+
+    /// If set, don't advance to the next glyph before going to the new state.
+    fn has_advance(&self) -> bool;
+
+    /// If set, push this glyph on the kerning stack. Used by format1 subtable.
+    fn has_push(&self) -> bool;
+
+    /// If set, remember this glyph as the marked glyph. Used by format4 subtable.
+    fn has_mark(&self) -> bool;
+
+    fn action_index(self) -> u16;
+    fn is_actionable(&self) -> bool;
+}
+
+impl KerxEntryDataExt for aat::ExtendedStateEntry<kerx::EntryData> {
+    fn new_state(&self) -> u16 { self.new_state }
+    fn flags(&self) -> u16 { self.flags }
+    fn has_reset(&self) -> bool { self.flags & 0x2000 != 0 }
+    fn has_advance(&self) -> bool { self.flags & 0x4000 == 0 }
+    fn has_push(&self) -> bool { self.flags & 0x8000 != 0 }
+    fn has_mark(&self) -> bool { self.flags & 0x8000 != 0 }
+    fn action_index(self) -> u16 { self.extra.action_index }
+    fn is_actionable(&self) -> bool { self.extra.action_index != 0xFFFF }
+}
+
+fn apply_state_machine_kerning<T, E>(
     coverage: kerx::Coverage,
     aat: &T,
     driver: &mut dyn StateTableDriver<T, E>,
     plan: &ShapePlan,
     buffer: &mut Buffer,
-) {
-    let mut state = aat::START_OF_TEXT;
+)
+where T: kerx::StateTableExt<E>,
+      E: FromData + Copy,
+      aat::ExtendedStateEntry<E>: KerxEntryDataExt
+{
+    let mut state = START_OF_TEXT;
     buffer.idx = 0;
     loop {
         let class = if buffer.idx < buffer.len {
@@ -172,20 +207,20 @@ fn apply_state_machine_kerning<E: aat::Entry, T: aat::StateTable2<E>>(
             aat::class::END_OF_TEXT
         };
 
-        let entry: E = match aat.entry(state, class) {
+        let entry: aat::ExtendedStateEntry<E> = match aat.entry(state, class) {
             Some(v) => v,
             None => break,
         };
 
         // Unsafe-to-break before this if not in state 0, as things might
         // go differently if we start from state 0 here.
-        if state != aat::START_OF_TEXT &&
+        if state != START_OF_TEXT &&
             buffer.backtrack_len() != 0 &&
             buffer.idx < buffer.len
         {
             // If there's no value and we're just epsilon-transitioning to state 0, safe to break.
             if   entry.is_actionable() ||
-                !(entry.new_state() == aat::START_OF_TEXT && !entry.has_advance())
+                !(entry.new_state() == START_OF_TEXT && !entry.has_advance())
             {
                 buffer.unsafe_to_break_from_outbuffer(buffer.backtrack_len() - 1, buffer.idx + 1);
             }
@@ -193,7 +228,7 @@ fn apply_state_machine_kerning<E: aat::Entry, T: aat::StateTable2<E>>(
 
         // Unsafe-to-break if end-of-text would kick in here.
         if buffer.idx + 2 <= buffer.len {
-            let end_entry: E = match aat.entry(state, aat::class::END_OF_TEXT) {
+            let end_entry: aat::ExtendedStateEntry<E> = match aat.entry(state, aat::class::END_OF_TEXT) {
                 Some(v) => v,
                 None => break,
             };
@@ -219,9 +254,9 @@ fn apply_state_machine_kerning<E: aat::Entry, T: aat::StateTable2<E>>(
 }
 
 
-trait StateTableDriver<Table, Entry> {
-    fn is_actionable(&self, entry: Entry) -> bool;
-    fn transition(&mut self, aat: &Table, entry: Entry,
+trait StateTableDriver<Table, E: FromData> {
+    fn is_actionable(&self, entry: aat::ExtendedStateEntry<E>) -> bool;
+    fn transition(&mut self, aat: &Table, entry: aat::ExtendedStateEntry<E>,
                   has_cross_stream: bool, plan: &ShapePlan, buffer: &mut Buffer) -> Option<()>;
 }
 
@@ -231,22 +266,19 @@ struct Driver1 {
     depth: usize,
 }
 
-impl StateTableDriver<kerx::format1::StateTable<'_>, kerx::format1::Entry> for Driver1 {
-    fn is_actionable(&self, entry: kerx::format1::Entry) -> bool {
-        use aat::Entry;
+impl StateTableDriver<kerx::format1::StateTable<'_>, kerx::EntryData> for Driver1 {
+    fn is_actionable(&self, entry: aat::ExtendedStateEntry<kerx::EntryData>) -> bool {
         entry.is_actionable()
     }
 
     fn transition(
         &mut self,
         aat: &kerx::format1::StateTable,
-        entry: kerx::format1::Entry,
+        entry: aat::ExtendedStateEntry<kerx::EntryData>,
         has_cross_stream: bool,
         plan: &ShapePlan,
         buffer: &mut Buffer,
     ) -> Option<()> {
-        use aat::Entry;
-
         if entry.has_reset() {
             self.depth = 0;
         }
@@ -263,7 +295,7 @@ impl StateTableDriver<kerx::format1::StateTable<'_>, kerx::format1::Entry> for D
         if entry.is_actionable() && self.depth != 0 {
             let tuple_count = u16::try_from(aat.tuple_count.max(1)).ok()?;
 
-            let mut action_index = entry.action_index;
+            let mut action_index = entry.action_index();
 
             // From Apple 'kern' spec:
             // "Each pops one glyph from the kerning stack and applies the kerning value to it.
@@ -342,26 +374,24 @@ struct Driver4<'a> {
     ankr_table: Option<ankr::Table<'a>>,
 }
 
-impl StateTableDriver<kerx::format4::StateTable<'_>, kerx::format4::Entry> for Driver4<'_> {
+impl StateTableDriver<kerx::format4::StateTable<'_>, kerx::EntryData> for Driver4<'_> {
     // TODO: remove
-    fn is_actionable(&self, entry: kerx::format4::Entry) -> bool {
-        use aat::Entry;
+    fn is_actionable(&self, entry: aat::ExtendedStateEntry<kerx::EntryData>) -> bool {
         entry.is_actionable()
     }
 
     fn transition(
         &mut self,
         aat: &kerx::format4::StateTable,
-        entry: kerx::format4::Entry,
+        entry: aat::ExtendedStateEntry<kerx::EntryData>,
         _has_cross_stream: bool,
         _opt: &ShapePlan,
         buffer: &mut Buffer,
     ) -> Option<()> {
-        use ttf_parser::parser::{Stream, FromData};
-        use aat::Entry;
+        use ttf_parser::parser::Stream;
 
         if self.mark_set && entry.is_actionable() && buffer.idx < buffer.len {
-            let points_data_offset = usize::from(entry.action_index) * u16::SIZE;
+            let points_data_offset = usize::from(entry.action_index()) * u16::SIZE;
             let mut s = Stream::new_at(aat.control_points_data, points_data_offset)?;
 
             // Note: I wasn't able to find any fonts that actually use