+++ src/plan.rs
@@ -91,7 +91,7 @@ impl<'a> ShapePlanner<'a> {
         let script_fallback_mark_positioning = shaper.fallback_position;
 
         // https://github.com/harfbuzz/harfbuzz/issues/2124
-        let apply_morx = face.morx.is_some() && (direction.is_horizontal() || !face.gsub.is_some());
+        let apply_morx = face.tables().morx.is_some() && (direction.is_horizontal() || !face.gsub.is_some());
 
         // https://github.com/harfbuzz/harfbuzz/issues/1528
         if apply_morx && shaper as *const _ != &DEFAULT_SHAPER as *const _ {