+++ src/tables/mod.rs
@@ -1,4 +1,2 @@
 pub mod kern;
 pub mod kerx;
-pub mod morx;
-pub mod aat;