+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
-- `GSUB`, `GPOS`, `ankr`, `feat` and `trak` tables parsing was moved to `ttf-parser`.
+- `GSUB`, `GPOS`, `ankr`, `feat`, `morx` and `trak` tables parsing was moved to `ttf-parser`.
 
 ### Fixed
 - Multiple issues in AAT Lookup Table parsing.